+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 ## [UNRELEASED] - (1.0.1)
 
 ### Fixed
-- Instead of looking for `latest.log` in the component log directory the cli client now picks the file with the newest
-changes in the directory.
+- Instead of looking for `latest.log` in the component log directory the ui clients now pick the file with the most
+recent file change.
 
 ### Added
 - Ability to display real information in the hosts stat columns.